move auth actions so they use serveraction hook
